Refactor streaming classes into more defined packages
