+++ sdl_android/src/androidTest/java/com/smartdevicelink/test/streaming/RTPH264PacketizerTest.java
@@ -35,7 +35,7 @@ import com.smartdevicelink.protocol.ProtocolMessage;
 import com.smartdevicelink.protocol.enums.SessionType;
 import com.smartdevicelink.proxy.interfaces.IVideoStreamListener;
 import com.smartdevicelink.streaming.IStreamListener;
-import com.smartdevicelink.streaming.RTPH264Packetizer;
+import com.smartdevicelink.streaming.video.RTPH264Packetizer;
 import com.smartdevicelink.transport.BTTransportConfig;
 
 import junit.framework.TestCase;
@@ -46,7 +46,7 @@ import java.nio.ByteBuffer;
 import java.nio.ByteOrder;
 
 /**
- * This class includes a unit test for {@link com.smartdevicelink.streaming.RTPH264Packetizer}.
+ * This class includes a unit test for {@link RTPH264Packetizer}.
  *
  * @author Sho Amano
  */
@@ -309,7 +309,7 @@ public class RTPH264PacketizerTest extends TestCase {
 	}
 
 	/**
-	 * Test for {@link com.smartdevicelink.streaming.RTPH264Packetizer#setPayloadType(byte)}
+	 * Test for {@link RTPH264Packetizer#setPayloadType(byte)}
 	 */
 	public void testSetPayloadType() {
 		byte pt = (byte)123;
@@ -340,7 +340,7 @@ public class RTPH264PacketizerTest extends TestCase {
 	}
 
 	/**
-	 * Test for {@link com.smartdevicelink.streaming.RTPH264Packetizer#setSSRC(int)}
+	 * Test for {@link RTPH264Packetizer#setSSRC(int)}
 	 */
 	public void testSetSSRC() {
 		int ssrc = 0xFEDCBA98;
@@ -372,8 +372,8 @@ public class RTPH264PacketizerTest extends TestCase {
 	}
 
 	/**
-	 * Test for {@link com.smartdevicelink.streaming.RTPH264Packetizer#pause()} and
-	 * {@link com.smartdevicelink.streaming.RTPH264Packetizer#resume()}
+	 * Test for {@link RTPH264Packetizer#pause()} and
+	 * {@link RTPH264Packetizer#resume()}
 	 */
 	public void testPauseResume() {
 		int index = 0;
@@ -443,7 +443,7 @@ public class RTPH264PacketizerTest extends TestCase {
 	}
 
 	/**
-	 * Test for {@link com.smartdevicelink.streaming.RTPH264Packetizer#sendFrame(byte[], int, int, long)}
+	 * Test for {@link RTPH264Packetizer#sendFrame(byte[], int, int, long)}
 	 */
 	public void testSendFrameInterfaceWithArray() {
 		StreamVerifier verifier = new StreamVerifier(SAMPLE_STREAM);
@@ -472,7 +472,7 @@ public class RTPH264PacketizerTest extends TestCase {
 	}
 
 	/**
-	 * Test for {@link com.smartdevicelink.streaming.RTPH264Packetizer#sendFrame(ByteBuffer, long)}
+	 * Test for {@link RTPH264Packetizer#sendFrame(ByteBuffer, long)}
 	 */
 	public void testSendFrameInterfaceWithByteBuffer() {
 		StreamVerifier verifier = new StreamVerifier(SAMPLE_STREAM);
@@ -501,7 +501,7 @@ public class RTPH264PacketizerTest extends TestCase {
 	}
 
 	/**
-	 * Test for {@link com.smartdevicelink.streaming.RTPH264Packetizer#sendFrame(ByteBuffer, long)}
+	 * Test for {@link RTPH264Packetizer#sendFrame(ByteBuffer, long)}
 	 * with direct ByteBuffer
 	 */
 	public void testSendFrameInterfaceWithDirectByteBuffer() {
