+++ sdl_android/src/main/java/com/smartdevicelink/streaming/audio/AudioStreamingParams.java
@@ -28,7 +28,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package com.smartdevicelink.streaming;
+package com.smartdevicelink.streaming.audio;
 
 /**
  * A struct to hold audio format information that are common to codecs.
